recentUnreadCount adjusts for latency
Pass a message_id into recentUnreadCount and ensure the count includes that id,
regardless of whether in the recentUnread query result set.  This allows brand
new messages to be included in the count that have not been indexed yet, since
the query uses a global secondary index, which is only eventually constistent.