query messages returns at least limit
Automatically rerun queries on the message collection to build up a result set
at least the limit (40) size. This prevents clients having to page through empty
result sets. Dynamo queries can be below the limit when there is a filter
expression, in this case handled attribute not exists indicating the upload has
not completed.