clean recorders table
Scan table for api key and put results into a channel. Recursively call scan
until there are no more results and add everything to the same channel, managing
LastEvaluatedKey and ExclusiveStartKey within the scan function. Pipe the scan
results channel to a discard routine one at a time. Could be improved with
batchWrite to delete up to 25 at a time.